Const volatile region list pointer
The region pointer (region_list) has to be const so it can't be changed
at runtime, but in addition, it has to be changed at runtime (in the
init routine) by mapping the corresponding memory region as writable.
Thus, to avoid optimization, the pointer has to be volatile too.
Since const volatile variables are placed in a region, that is not
included in dl_iterate_phdr, we need an additional abstraction layer,
i.e. a const volatile pointer to the region pointer.