gcc/fpprotect.c: store addr of global fp in fps, also struct initialization
Each function address that is used as a func pointer is stored in a global variable.
In the func pointer itself, a pointer to that global var is stored and it is derefenced
twice when used. This is needed to be able to initialize global const vars in a global
constructor.